# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Base Api",
        "description": "API do projeto Base",
        "contact": {
            "email": "email@test.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Base API"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realizar autenticação",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cpf",
                                    "password"
                                ],
                                "properties": {
                                    "cpf": {
                                        "type": "string",
                                        "example": "00000000191"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "teste123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "O servidor não entende o tipo de conteúdo da entidade de solicitação",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Ocoreu um erro"
                    },
                    "404": {
                        "description": "Página não localizada"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realiza Logout",
                "operationId": "logout",
                "responses": {
                    "201": {
                        "description": "Logout realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Logout realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "O servidor não entende o tipo de conteúdo da entidade de solicitação",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Ocoreu um erro"
                    },
                    "404": {
                        "description": "Página não localizada"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/refresh": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realiza RefreshToken",
                "operationId": "refresh",
                "responses": {
                    "201": {
                        "description": "RefreshToken realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "RefreshToken realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "O servidor não entende o tipo de conteúdo da entidade de solicitação",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Ocoreu um erro"
                    },
                    "404": {
                        "description": "Página não localizada"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/reset-password": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Resetar senha - É necessário o token enviado por e-mail",
                "operationId": "reset-password",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cpf",
                                    "token",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "cpf": {
                                        "type": "string",
                                        "example": "00000000191"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "token-email"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "teste@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "teste123456"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "teste123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login recuperado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Login recuperado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "O servidor não entende o tipo de conteúdo da entidade de solicitação",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Ocoreu um erro"
                    },
                    "404": {
                        "description": "Página não localizada"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/forgot-password": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Envia o e-mail de recuperação com o token de recuperação",
                "operationId": "forgot-password",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "teste@mail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login recuperado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Login recuperado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "O servidor não entende o tipo de conteúdo da entidade de solicitação",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Ocoreu um erro"
                    },
                    "404": {
                        "description": "Página não localizada"
                    }
                }
            }
        },
        "/verify-email/{id}/{hash}": {
            "get": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Confirmação de Verificação de E-mail",
                "operationId": "verify-email",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "00000000191"
                        }
                    },
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "hash",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": ""
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Checagem de e-mail realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Checagem de e-mail realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "O servidor não entende o tipo de conteúdo da entidade de solicitação",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Ocoreu um erro"
                    },
                    "404": {
                        "description": "Página não localizada"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "in": "header",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "/api/auth/auth",
                        "scopes": {}
                    }
                },
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}